[UPDATE] Change TopicPage to PracticePage
